+++ static/js/dropzone.js
@@ -1,0 +1,454 @@
+<!DOCTYPE html>
+<html class="" lang="en">
+<head prefix="og: http://ogp.me/ns#">
+<meta charset="utf-8">
+<meta content="IE=edge" http-equiv="X-UA-Compatible">
+<meta content="object" property="og:type">
+<meta content="GitLab" property="og:site_name">
+<meta content="Artifacts · release (#78454225) · Jobs · Matias Meno / Dropzone" property="og:title">
+<meta content="Dropzone is an easy to use drag&#39;n&#39;drop library. It supports image previews and shows nice progress bars. http://www.dropzonejs.com" property="og:description">
+<meta content="/uploads/-/system/project/avatar/2712276/dropzone-logo.png" property="og:image">
+<meta content="64" property="og:image:width">
+<meta content="64" property="og:image:height">
+<meta content="https://gitlab.com/meno/dropzone/-/jobs/78454225/artifacts/file/dist/dropzone.js" property="og:url">
+<meta content="summary" property="twitter:card">
+<meta content="Artifacts · release (#78454225) · Jobs · Matias Meno / Dropzone" property="twitter:title">
+<meta content="Dropzone is an easy to use drag&#39;n&#39;drop library. It supports image previews and shows nice progress bars. http://www.dropzonejs.com" property="twitter:description">
+<meta content="/uploads/-/system/project/avatar/2712276/dropzone-logo.png" property="twitter:image">
+
+<title>Artifacts · release (#78454225) · Jobs · Matias Meno / Dropzone · GitLab</title>
+<meta content="Dropzone is an easy to use drag&#39;n&#39;drop library. It supports image previews and shows nice progress bars. http://www.dropzonejs.com" name="description">
+<link rel="shortcut icon" type="image/png" href="https://gitlab.com/assets/favicon-7901bd695fb93edb07975966062049829afb56cf11511236e61bcf425070e36e.png" id="favicon" data-original-href="https://gitlab.com/assets/favicon-7901bd695fb93edb07975966062049829afb56cf11511236e61bcf425070e36e.png" />
+<link rel="stylesheet" media="all" href="https://assets.gitlab-static.net/assets/application-1d86daecfb6704a29ef0e9f42ab3478baed22a68f8e1f8aa719ec7a8698d2129.css" />
+<link rel="stylesheet" media="print" href="https://assets.gitlab-static.net/assets/print-c8ff536271f8974b8a9a5f75c0ca25d2b8c1dceb4cff3c01d1603862a0bdcbfc.css" />
+
+
+<script>
+//<![CDATA[
+window.gon={};gon.api_version="v4";gon.default_avatar_url="https://assets.gitlab-static.net/assets/no_avatar-849f9c04a3a0d0cea2424ae97b27447dc64a7dbfae83c036c45b403392f0e8ba.png";gon.max_file_size=10;gon.asset_host="https://assets.gitlab-static.net";gon.webpack_public_path="https://assets.gitlab-static.net/assets/webpack/";gon.relative_url_root="";gon.shortcuts_path="/help/shortcuts";gon.user_color_scheme="white";gon.sentry_dsn="https://526a2f38a53d44e3a8e69bfa001d1e8b@sentry.gitlap.com/15";gon.gitlab_url="https://gitlab.com";gon.revision="aedb3c9";gon.gitlab_logo="https://assets.gitlab-static.net/assets/gitlab_logo-7ae504fe4f68fdebb3c2034e36621930cd36ea87924c11ff65dbcb8ed50dca58.png";gon.sprite_icons="https://gitlab.com/assets/icons-287b331848727a2f75ca16fc538c5a50deacfed69bdd96c6e59f1d26b422f961.svg";gon.sprite_file_icons="https://gitlab.com/assets/file_icons-7262fc6897e02f1ceaf8de43dc33afa5e4f9a2067f4f68ef77dcc87946575e9e.svg";gon.emoji_sprites_css_path="https://assets.gitlab-static.net/assets/emoji_sprites-289eccffb1183c188b630297431be837765d9ff4aed6130cf738586fb307c170.css";gon.test_env=false;gon.suggested_label_colors=["#0033CC","#428BCA","#44AD8E","#A8D695","#5CB85C","#69D100","#004E00","#34495E","#7F8C8D","#A295D6","#5843AD","#8E44AD","#FFECDB","#AD4363","#D10069","#CC0033","#FF0000","#D9534F","#D1D100","#F0AD4E","#AD8D43"];
+//]]>
+</script>
+
+<script src="https://assets.gitlab-static.net/assets/webpack/runtime.754cd766.bundle.js" defer="defer"></script>
+<script src="https://assets.gitlab-static.net/assets/webpack/main.1fed6912.chunk.js" defer="defer"></script>
+<script src="https://assets.gitlab-static.net/assets/webpack/raven.d2dcfa91.chunk.js" defer="defer"></script>
+<script src="https://assets.gitlab-static.net/assets/webpack/pages.projects.artifacts.file.9e47d86b.chunk.js" defer="defer"></script>
+
+<meta name="csrf-param" content="authenticity_token" />
+<meta name="csrf-token" content="SjIRRcxc1Xo6240roZtDHd/ygEc+aM9wXcbYv/90gqA6TqSqYUwSUGLaF19+TQCMetIZGhVmu19infIoLvaZ9Q==" />
+<meta content="origin-when-cross-origin" name="referrer">
+<meta content="width=device-width, initial-scale=1, maximum-scale=1" name="viewport">
+<meta content="#474D57" name="theme-color">
+<link rel="apple-touch-icon" type="image/x-icon" href="https://assets.gitlab-static.net/assets/touch-icon-iphone-5a9cee0e8a51212e70b90c87c12f382c428870c0ff67d1eb034d884b78d2dae7.png" />
+<link rel="apple-touch-icon" type="image/x-icon" href="https://assets.gitlab-static.net/assets/touch-icon-ipad-a6eec6aeb9da138e507593b464fdac213047e49d3093fc30e90d9a995df83ba3.png" sizes="76x76" />
+<link rel="apple-touch-icon" type="image/x-icon" href="https://assets.gitlab-static.net/assets/touch-icon-iphone-retina-72e2aadf86513a56e050e7f0f2355deaa19cc17ed97bbe5147847f2748e5a3e3.png" sizes="120x120" />
+<link rel="apple-touch-icon" type="image/x-icon" href="https://assets.gitlab-static.net/assets/touch-icon-ipad-retina-8ebe416f5313483d9c1bc772b5bbe03ecad52a54eba443e5215a22caed2a16a2.png" sizes="152x152" />
+<link color="rgb(226, 67, 41)" href="https://assets.gitlab-static.net/assets/logo-d36b5212042cebc89b96df4bf6ac24e43db316143e89926c0db839ff694d2de4.svg" rel="mask-icon">
+<meta content="https://assets.gitlab-static.net/assets/msapplication-tile-1196ec67452f618d39cdd85e2e3a542f76574c071051ae7effbfde01710eb17d.png" name="msapplication-TileImage">
+<meta content="#30353E" name="msapplication-TileColor">
+
+
+
+</head>
+
+<body class="ui-indigo " data-find-file="/meno/dropzone/find_file/master" data-group="" data-page="projects:artifacts:file" data-project="dropzone">
+
+
+
+<header class="navbar navbar-gitlab navbar-gitlab-new qa-navbar navbar-expand-sm">
+<a class="sr-only gl-accessibility" href="#content-body" tabindex="1">Skip to content</a>
+<div class="container-fluid">
+<div class="header-content">
+<div class="title-container">
+<h1 class="title">
+<a title="Dashboard" id="logo" href="/"><svg width="24" height="24" class="tanuki-logo" viewBox="0 0 36 36">
+  <path class="tanuki-shape tanuki-left-ear" fill="#e24329" d="M2 14l9.38 9v-9l-4-12.28c-.205-.632-1.176-.632-1.38 0z"/>
+  <path class="tanuki-shape tanuki-right-ear" fill="#e24329" d="M34 14l-9.38 9v-9l4-12.28c.205-.632 1.176-.632 1.38 0z"/>
+  <path class="tanuki-shape tanuki-nose" fill="#e24329" d="M18,34.38 3,14 33,14 Z"/>
+  <path class="tanuki-shape tanuki-left-eye" fill="#fc6d26" d="M18,34.38 11.38,14 2,14 6,25Z"/>
+  <path class="tanuki-shape tanuki-right-eye" fill="#fc6d26" d="M18,34.38 24.62,14 34,14 30,25Z"/>
+  <path class="tanuki-shape tanuki-left-cheek" fill="#fca326" d="M2 14L.1 20.16c-.18.565 0 1.2.5 1.56l17.42 12.66z"/>
+  <path class="tanuki-shape tanuki-right-cheek" fill="#fca326" d="M34 14l1.9 6.16c.18.565 0 1.2-.5 1.56L18 34.38z"/>
+</svg>
+
+<span class="logo-text d-none d-sm-block">
+<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 617 169"><path d="M315.26 2.97h-21.8l.1 162.5h88.3v-20.1h-66.5l-.1-142.4M465.89 136.95c-5.5 5.7-14.6 11.4-27 11.4-16.6 0-23.3-8.2-23.3-18.9 0-16.1 11.2-23.8 35-23.8 4.5 0 11.7.5 15.4 1.2v30.1h-.1m-22.6-98.5c-17.6 0-33.8 6.2-46.4 16.7l7.7 13.4c8.9-5.2 19.8-10.4 35.5-10.4 17.9 0 25.8 9.2 25.8 24.6v7.9c-3.5-.7-10.7-1.2-15.1-1.2-38.2 0-57.6 13.4-57.6 41.4 0 25.1 15.4 37.7 38.7 37.7 15.7 0 30.8-7.2 36-18.9l4 15.9h15.4v-83.2c-.1-26.3-11.5-43.9-44-43.9M557.63 149.1c-8.2 0-15.4-1-20.8-3.5V70.5c7.4-6.2 16.6-10.7 28.3-10.7 21.1 0 29.2 14.9 29.2 39 0 34.2-13.1 50.3-36.7 50.3m9.2-110.6c-19.5 0-30 13.3-30 13.3v-21l-.1-27.8h-21.3l.1 158.5c10.7 4.5 25.3 6.9 41.2 6.9 40.7 0 60.3-26 60.3-70.9-.1-35.5-18.2-59-50.2-59M77.9 20.6c19.3 0 31.8 6.4 39.9 12.9l9.4-16.3C114.5 6 97.3 0 78.9 0 32.5 0 0 28.3 0 85.4c0 59.8 35.1 83.1 75.2 83.1 20.1 0 37.2-4.7 48.4-9.4l-.5-63.9V75.1H63.6v20.1h38l.5 48.5c-5 2.5-13.6 4.5-25.3 4.5-32.2 0-53.8-20.3-53.8-63-.1-43.5 22.2-64.6 54.9-64.6M231.43 2.95h-21.3l.1 27.3v94.3c0 26.3 11.4 43.9 43.9 43.9 4.5 0 8.9-.4 13.1-1.2v-19.1c-3.1.5-6.4.7-9.9.7-17.9 0-25.8-9.2-25.8-24.6v-65h35.7v-17.8h-35.7l-.1-38.5M155.96 165.47h21.3v-124h-21.3v124M155.96 24.37h21.3V3.07h-21.3v21.3"/></svg>
+
+</span>
+</a></h1>
+<ul class="list-unstyled navbar-sub-nav">
+<li class="home"><a title="Projects" class="dashboard-shortcuts-projects" href="/explore">Projects
+</a></li><li class=""><a title="Groups" class="dashboard-shortcuts-groups" href="/explore/groups">Groups
+</a></li><li class=""><a title="Snippets" class="dashboard-shortcuts-snippets" href="/explore/snippets">Snippets
+</a></li><li>
+<a title="About GitLab CE" href="/help">Help</a>
+</li>
+</ul>
+
+</div>
+<div class="navbar-collapse collapse">
+<ul class="nav navbar-nav">
+<li class="nav-item d-none d-sm-none d-md-block">
+<div class="has-location-badge search search-form">
+<form class="form-inline" action="/search" accept-charset="UTF-8" method="get"><input name="utf8" type="hidden" value="&#x2713;" /><div class="search-input-container">
+<div class="location-badge">This project</div>
+<div class="search-input-wrap">
+<div class="dropdown" data-url="/search/autocomplete">
+<input type="search" name="search" id="search" placeholder="Search" class="search-input dropdown-menu-toggle no-outline js-search-dashboard-options" spellcheck="false" tabindex="1" autocomplete="off" data-issues-path="/dashboard/issues" data-mr-path="/dashboard/merge_requests" aria-label="Search" />
+<button class="hidden js-dropdown-search-toggle" data-toggle="dropdown" type="button"></button>
+<div class="dropdown-menu dropdown-select">
+<div class="dropdown-content"><ul>
+<li class="dropdown-menu-empty-item">
+<a>
+Loading...
+</a>
+</li>
+</ul>
+</div><div class="dropdown-loading"><i aria-hidden="true" data-hidden="true" class="fa fa-spinner fa-spin"></i></div>
+</div>
+<svg class="s16 search-icon"><use xlink:href="https://gitlab.com/assets/icons-287b331848727a2f75ca16fc538c5a50deacfed69bdd96c6e59f1d26b422f961.svg#search"></use></svg>
+<svg class="s16 clear-icon js-clear-input"><use xlink:href="https://gitlab.com/assets/icons-287b331848727a2f75ca16fc538c5a50deacfed69bdd96c6e59f1d26b422f961.svg#close"></use></svg>
+</div>
+</div>
+</div>
+<input type="hidden" name="group_id" id="group_id" class="js-search-group-options" />
+<input type="hidden" name="project_id" id="search_project_id" value="2712276" class="js-search-project-options" data-project-path="dropzone" data-name="Dropzone" data-issues-path="/meno/dropzone/issues" data-mr-path="/meno/dropzone/merge_requests" data-issues-disabled="false" />
+<input type="hidden" name="search_code" id="search_code" value="true" />
+<input type="hidden" name="repository_ref" id="repository_ref" />
+
+<div class="search-autocomplete-opts hide" data-autocomplete-path="/search/autocomplete" data-autocomplete-project-id="2712276"></div>
+</form></div>
+
+</li>
+<li class="nav-item d-inline-block d-sm-none d-md-none">
+<a title="Search" aria-label="Search" data-toggle="tooltip" data-placement="bottom" data-container="body" href="/search"><svg class="s16"><use xlink:href="https://gitlab.com/assets/icons-287b331848727a2f75ca16fc538c5a50deacfed69bdd96c6e59f1d26b422f961.svg#search"></use></svg>
+</a></li>
+<li class="nav-item">
+<div>
+<a class="btn btn-sign-in" href="/users/sign_in?redirect_to_referer=yes">Sign in / Register</a>
+</div>
+</li>
+</ul>
+</div>
+<button class="navbar-toggler d-block d-sm-none" type="button">
+<span class="sr-only">Toggle navigation</span>
+<svg class="s12 more-icon js-navbar-toggle-right"><use xlink:href="https://gitlab.com/assets/icons-287b331848727a2f75ca16fc538c5a50deacfed69bdd96c6e59f1d26b422f961.svg#more"></use></svg>
+<svg class="s12 close-icon js-navbar-toggle-left"><use xlink:href="https://gitlab.com/assets/icons-287b331848727a2f75ca16fc538c5a50deacfed69bdd96c6e59f1d26b422f961.svg#close"></use></svg>
+</button>
+</div>
+</div>
+</header>
+
+<div class="layout-page page-with-contextual-sidebar">
+<div class="nav-sidebar">
+<div class="nav-sidebar-inner-scroll">
+<div class="context-header">
+<a title="Dropzone" href="/meno/dropzone"><div class="avatar-container s40 project-avatar">
+<img alt="Dropzone" class="avatar s40 avatar-tile lazy" data-src="https://assets.gitlab-static.net/uploads/-/system/project/avatar/2712276/dropzone-logo.png" src="data:image/gif;base64,R0lGODlhAQABAAAAACH5BAEKAAEALAAAAAABAAEAAAICTAEAOw==" />
+</div>
+<div class="sidebar-context-title">
+Dropzone
+</div>
+</a></div>
+<ul class="sidebar-top-level-items">
+<li class="home"><a class="shortcuts-project" href="/meno/dropzone"><div class="nav-icon-container">
+<svg><use xlink:href="https://gitlab.com/assets/icons-287b331848727a2f75ca16fc538c5a50deacfed69bdd96c6e59f1d26b422f961.svg#project"></use></svg>
+</div>
+<span class="nav-item-name">
+Project
+</span>
+</a><ul class="sidebar-sub-level-items">
+<li class="fly-out-top-item"><a href="/meno/dropzone"><strong class="fly-out-top-item-name">
+Project
+</strong>
+</a></li><li class="divider fly-out-top-item"></li>
+<li class=""><a title="Project details" class="shortcuts-project" href="/meno/dropzone"><span>Details</span>
+</a></li><li class=""><a title="Activity" class="shortcuts-project-activity" href="/meno/dropzone/activity"><span>Activity</span>
+</a></li>
+<li class=""><a title="Cycle Analytics" class="shortcuts-project-cycle-analytics" href="/meno/dropzone/cycle_analytics"><span>Cycle Analytics</span>
+</a></li></ul>
+</li><li class=""><a class="shortcuts-tree" href="/meno/dropzone/tree/master"><div class="nav-icon-container">
+<svg><use xlink:href="https://gitlab.com/assets/icons-287b331848727a2f75ca16fc538c5a50deacfed69bdd96c6e59f1d26b422f961.svg#doc_text"></use></svg>
+</div>
+<span class="nav-item-name">
+Repository
+</span>
+</a><ul class="sidebar-sub-level-items">
+<li class="fly-out-top-item"><a href="/meno/dropzone/tree/master"><strong class="fly-out-top-item-name">
+Repository
+</strong>
+</a></li><li class="divider fly-out-top-item"></li>
+<li class=""><a href="/meno/dropzone/tree/master">Files
+</a></li><li class=""><a href="/meno/dropzone/commits/master">Commits
+</a></li><li class=""><a href="/meno/dropzone/branches">Branches
+</a></li><li class=""><a href="/meno/dropzone/tags">Tags
+</a></li><li class=""><a href="/meno/dropzone/graphs/master">Contributors
+</a></li><li class=""><a href="/meno/dropzone/network/master">Graph
+</a></li><li class=""><a href="/meno/dropzone/compare?from=master&amp;to=master">Compare
+</a></li><li class=""><a href="/meno/dropzone/graphs/master/charts">Charts
+</a></li><li class=""><a href="/meno/dropzone/path_locks">Locked Files
+</a></li>
+</ul>
+</li><li class=""><a class="shortcuts-issues" href="/meno/dropzone/issues"><div class="nav-icon-container">
+<svg><use xlink:href="https://gitlab.com/assets/icons-287b331848727a2f75ca16fc538c5a50deacfed69bdd96c6e59f1d26b422f961.svg#issues"></use></svg>
+</div>
+<span class="nav-item-name">
+Issues
+</span>
+<span class="badge badge-pill count issue_counter">
+77
+</span>
+</a><ul class="sidebar-sub-level-items">
+<li class="fly-out-top-item"><a href="/meno/dropzone/issues"><strong class="fly-out-top-item-name">
+Issues
+</strong>
+<span class="badge badge-pill count issue_counter fly-out-badge">
+77
+</span>
+</a></li><li class="divider fly-out-top-item"></li>
+<li class=""><a title="Issues" href="/meno/dropzone/issues"><span>
+List
+</span>
+</a></li><li class=""><a title="Boards" href="/meno/dropzone/boards"><span>
+Boards
+</span>
+</a></li><li class=""><a title="Labels" href="/meno/dropzone/labels"><span>
+Labels
+</span>
+</a></li><li class=""><a title="Service Desk" href="/meno/dropzone/issues/service_desk">Service Desk
+</a></li>
+<li class=""><a title="Milestones" href="/meno/dropzone/milestones"><span>
+Milestones
+</span>
+</a></li></ul>
+</li><li class=""><a class="shortcuts-merge_requests" href="/meno/dropzone/merge_requests"><div class="nav-icon-container">
+<svg><use xlink:href="https://gitlab.com/assets/icons-287b331848727a2f75ca16fc538c5a50deacfed69bdd96c6e59f1d26b422f961.svg#git-merge"></use></svg>
+</div>
+<span class="nav-item-name">
+Merge Requests
+</span>
+<span class="badge badge-pill count merge_counter js-merge-counter">
+13
+</span>
+</a><ul class="sidebar-sub-level-items is-fly-out-only">
+<li class="fly-out-top-item"><a href="/meno/dropzone/merge_requests"><strong class="fly-out-top-item-name">
+Merge Requests
+</strong>
+<span class="badge badge-pill count merge_counter js-merge-counter fly-out-badge">
+13
+</span>
+</a></li></ul>
+</li><li class="active"><a class="shortcuts-pipelines" href="/meno/dropzone/pipelines"><div class="nav-icon-container">
+<svg><use xlink:href="https://gitlab.com/assets/icons-287b331848727a2f75ca16fc538c5a50deacfed69bdd96c6e59f1d26b422f961.svg#rocket"></use></svg>
+</div>
+<span class="nav-item-name">
+CI / CD
+</span>
+</a><ul class="sidebar-sub-level-items">
+<li class="fly-out-top-item active"><a href="/meno/dropzone/pipelines"><strong class="fly-out-top-item-name">
+CI / CD
+</strong>
+</a></li><li class="divider fly-out-top-item"></li>
+<li class=""><a title="Pipelines" class="shortcuts-pipelines" href="/meno/dropzone/pipelines"><span>
+Pipelines
+</span>
+</a></li><li class="active"><a title="Jobs" class="shortcuts-builds" href="/meno/dropzone/-/jobs"><span>
+Jobs
+</span>
+</a></li><li class=""><a title="Schedules" class="shortcuts-builds" href="/meno/dropzone/pipeline_schedules"><span>
+Schedules
+</span>
+</a></li><li class=""><a title="Charts" class="shortcuts-pipelines-charts" href="/meno/dropzone/pipelines/charts"><span>
+Charts
+</span>
+</a></li></ul>
+</li><li class=""><a class="shortcuts-container-registry" href="/meno/dropzone/container_registry"><div class="nav-icon-container">
+<svg><use xlink:href="https://gitlab.com/assets/icons-287b331848727a2f75ca16fc538c5a50deacfed69bdd96c6e59f1d26b422f961.svg#disk"></use></svg>
+</div>
+<span class="nav-item-name">
+Registry
+</span>
+</a><ul class="sidebar-sub-level-items is-fly-out-only">
+<li class="fly-out-top-item"><a href="/meno/dropzone/container_registry"><strong class="fly-out-top-item-name">
+Registry
+</strong>
+</a></li></ul>
+</li><li class=""><a class="shortcuts-wiki" href="/meno/dropzone/wikis/home"><div class="nav-icon-container">
+<svg><use xlink:href="https://gitlab.com/assets/icons-287b331848727a2f75ca16fc538c5a50deacfed69bdd96c6e59f1d26b422f961.svg#book"></use></svg>
+</div>
+<span class="nav-item-name">
+Wiki
+</span>
+</a><ul class="sidebar-sub-level-items is-fly-out-only">
+<li class="fly-out-top-item"><a href="/meno/dropzone/wikis/home"><strong class="fly-out-top-item-name">
+Wiki
+</strong>
+</a></li></ul>
+</li><li class=""><a title="Members" class="shortcuts-tree" href="/meno/dropzone/settings/members"><div class="nav-icon-container">
+<svg><use xlink:href="https://gitlab.com/assets/icons-287b331848727a2f75ca16fc538c5a50deacfed69bdd96c6e59f1d26b422f961.svg#users"></use></svg>
+</div>
+<span class="nav-item-name">
+Members
+</span>
+</a><ul class="sidebar-sub-level-items is-fly-out-only">
+<li class="fly-out-top-item"><a href="/meno/dropzone/project_members"><strong class="fly-out-top-item-name">
+Members
+</strong>
+</a></li></ul>
+</li><a class="toggle-sidebar-button js-toggle-sidebar" role="button" title="Toggle sidebar" type="button">
+<svg class=" icon-angle-double-left"><use xlink:href="https://gitlab.com/assets/icons-287b331848727a2f75ca16fc538c5a50deacfed69bdd96c6e59f1d26b422f961.svg#angle-double-left"></use></svg>
+<svg class=" icon-angle-double-right"><use xlink:href="https://gitlab.com/assets/icons-287b331848727a2f75ca16fc538c5a50deacfed69bdd96c6e59f1d26b422f961.svg#angle-double-right"></use></svg>
+<span class="collapse-text">Collapse sidebar</span>
+</a>
+<button name="button" type="button" class="close-nav-button"><svg class="s16"><use xlink:href="https://gitlab.com/assets/icons-287b331848727a2f75ca16fc538c5a50deacfed69bdd96c6e59f1d26b422f961.svg#close"></use></svg>
+<span class="collapse-text">Close sidebar</span>
+</button>
+<li class="hidden">
+<a title="Activity" class="shortcuts-project-activity" href="/meno/dropzone/activity"><span>
+Activity
+</span>
+</a></li>
+<li class="hidden">
+<a title="Network" class="shortcuts-network" href="/meno/dropzone/network/master">Graph
+</a></li>
+<li class="hidden">
+<a title="Charts" class="shortcuts-repository-charts" href="/meno/dropzone/graphs/master/charts">Charts
+</a></li>
+<li class="hidden">
+<a class="shortcuts-new-issue" href="/meno/dropzone/issues/new">Create a new issue
+</a></li>
+<li class="hidden">
+<a title="Jobs" class="shortcuts-builds" href="/meno/dropzone/-/jobs">Jobs
+</a></li>
+<li class="hidden">
+<a title="Commits" class="shortcuts-commits" href="/meno/dropzone/commits/master">Commits
+</a></li>
+<li class="hidden">
+<a title="Issue Boards" class="shortcuts-issue-boards" href="/meno/dropzone/boards">Issue Boards</a>
+</li>
+</ul>
+</div>
+</div>
+
+<div class="content-wrapper">
+
+<div class="mobile-overlay"></div>
+<div class="alert-wrapper">
+
+
+
+
+
+<nav class="breadcrumbs container-fluid container-limited" role="navigation">
+<div class="breadcrumbs-container">
+<button name="button" type="button" class="toggle-mobile-nav"><span class="sr-only">Open sidebar</span>
+<i aria-hidden="true" data-hidden="true" class="fa fa-bars"></i>
+</button><div class="breadcrumbs-links js-title-container">
+<ul class="list-unstyled breadcrumbs-list js-breadcrumbs-list">
+<li><a href="/meno">Matias Meno</a><svg class="s8 breadcrumbs-list-angle"><use xlink:href="https://gitlab.com/assets/icons-287b331848727a2f75ca16fc538c5a50deacfed69bdd96c6e59f1d26b422f961.svg#angle-right"></use></svg></li> <li><a href="/meno/dropzone"><img alt="Dropzone" class="avatar-tile lazy" width="15" height="15" data-src="https://assets.gitlab-static.net/uploads/-/system/project/avatar/2712276/dropzone-logo.png" src="data:image/gif;base64,R0lGODlhAQABAAAAACH5BAEKAAEALAAAAAABAAEAAAICTAEAOw==" /><span class="breadcrumb-item-text js-breadcrumb-item-text">Dropzone</span></a><svg class="s8 breadcrumbs-list-angle"><use xlink:href="https://gitlab.com/assets/icons-287b331848727a2f75ca16fc538c5a50deacfed69bdd96c6e59f1d26b422f961.svg#angle-right"></use></svg></li>
+
+<li>
+<h2 class="breadcrumbs-sub-title"><a href="/meno/dropzone/-/jobs/78454225/artifacts/file/dist/dropzone.js">Jobs</a></h2>
+</li>
+</ul>
+</div>
+
+</div>
+</nav>
+
+<div class="flash-container flash-container-page">
+</div>
+
+</div>
+<div class="container-fluid container-limited ">
+<div class="content" id="content-body">
+<div class="content-block build-header top-area page-content-header">
+<div class="header-content">
+<span class="ci-status ci-success has-tooltip" data-html title="Success">
+<svg><use xlink:href="https://gitlab.com/assets/icons-287b331848727a2f75ca16fc538c5a50deacfed69bdd96c6e59f1d26b422f961.svg#status_success"></use></svg>
+passed
+</span>
+
+<strong>
+Job
+<a class="js-build-id" href="/meno/dropzone/-/jobs/78454225">#78454225</a>
+</strong>
+in pipeline
+<strong>
+<a href="/meno/dropzone/pipelines/24858979">#24858979</a>
+</strong>
+for
+<strong>
+<a class="commit-sha" href="/meno/dropzone/commit/1df7f44c20575ae74633ed5eb6a2d11ad8234b25">1df7f44c</a>
+</strong>
+from
+<strong>
+<a class="ref-name" href="/meno/dropzone/commits/master">master</a>
+</strong>
+by
+<a href="/meno">
+<span class="d-none d-sm-inline">
+<img class="avatar s24 lazy" data-src="https://secure.gravatar.com/avatar/ce44e57e4bc9b51f9692e17bae6319e7?s=48&amp;d=identicon" src="data:image/gif;base64,R0lGODlhAQABAAAAACH5BAEKAAEALAAAAAABAAEAAAICTAEAOw==" />
+<strong data-placement="top" data-title="@meno" data-toggle="tooltip">
+Matias Meno
+</strong>
+</span>
+<strong class="d-inline d-sm-none">@meno</strong>
+</a>
+
+<time class="js-timeago" title="Jun 29, 2018 7:58pm" datetime="2018-06-29T19:58:42Z" data-toggle="tooltip" data-placement="top" data-container="body">Jun 29, 2018</time>
+</div>
+</div>
+
+<div class="tree-holder">
+<div class="nav-block">
+<ul class="breadcrumb repo-breadcrumb">
+<li class="breadcrumb-item">
+<a href="/meno/dropzone/-/jobs/78454225/artifacts/browse">Artifacts</a>
+</li>
+</ul>
+</div>
+<article class="file-holder">
+<div class="js-file-title file-title-flex-parent">
+<div class="file-header-content">
+<i aria-hidden="true" data-hidden="true" class="fa fa-file-text-o fa-fw"></i>
+<strong class="file-title-name">
+dropzone.js
+</strong>
+<button class="btn btn-clipboard btn-transparent prepend-left-5" data-toggle="tooltip" data-placement="bottom" data-container="body" data-class="btn-clipboard btn-transparent prepend-left-5" data-title="Copy file path to clipboard" data-clipboard-text="{&quot;text&quot;:&quot;dist/dropzone.js&quot;,&quot;gfm&quot;:&quot;`dist/dropzone.js`&quot;}" type="button" title="Copy file path to clipboard" aria-label="Copy file path to clipboard"><i aria-hidden="true" aria-hidden="true" data-hidden="true" class="fa fa-clipboard"></i></button>
+<small>
+118 KB
+</small>
+</div>
+
+<div class="file-actions d-none d-sm-block">
+
+<div class="btn-group" role="group"><a class="btn btn-sm has-tooltip" target="_blank" rel="noopener noreferrer" title="Download" data-container="body" href="/meno/dropzone/-/jobs/78454225/artifacts/raw/dist/dropzone.js"><svg><use xlink:href="https://gitlab.com/assets/icons-287b331848727a2f75ca16fc538c5a50deacfed69bdd96c6e59f1d26b422f961.svg#download"></use></svg></a></div>
+</div>
+</div>
+<div class="blob-viewer" data-type="simple">
+<div class="file-content code">
+<div class="nothing-here-block">
+The source could not be displayed because it is stored as a job artifact.
+You can
+<a target="_blank" rel="noopener noreferrer" href="/meno/dropzone/-/jobs/78454225/artifacts/raw/dist/dropzone.js">download it</a>
+instead.
+</div>
+</div>
+
+</div>
+
+
+</article>
+</div>
+
+</div>
+</div>
+</div>
+</div>
+
+
+
+</body>
+</html>
+